http-api/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "An HTTP API"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = {extras = ["async"], version = "^3.1.0"}
sqlalchemy = "^2.0.38"
alembic = "^1.14.1"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
aiohttp = "^3.11.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
